+++ src/main/java/com/lovetropics/minigames/common/dimension/DimensionUtils.java
@@ -1,7 +1,7 @@
 package com.lovetropics.minigames.common.dimension;
 
 import com.lovetropics.minigames.LoveTropics;
-import com.lovetropics.minigames.common.minigames.dimensions.CollaborativeConservationDimension;
+import com.lovetropics.minigames.common.minigames.dimensions.ConservationExplorationDimension;
 import com.lovetropics.minigames.common.minigames.dimensions.SurviveTheTideDimension;
 import com.tterrag.registrate.Registrate;
 import com.tterrag.registrate.util.entry.RegistryEntry;
@@ -25,8 +25,8 @@ public class DimensionUtils {
 			.dimensionTypeCallback(t -> SURVIVE_THE_TIDE_DIMENSION = t)
 			.register();
 
-	public static final RegistryEntry<ModDimension> COLLABORATIVE_CONSERVATION_MOD_DIMENSION = REGISTRATE
-			.dimension("collaborative_conservation_game", CollaborativeConservationDimension::new)
+	public static final RegistryEntry<ModDimension> CONSERVATION_EXPLORATION_MOD_DIMENSION = REGISTRATE
+			.dimension("conservation_exploration_game", ConservationExplorationDimension::new)
 			.keepLoaded(false)
 			.hasSkyLight(true)
 			.register();
