+++ src/main/java/com/lovetropics/minigames/common/minigames/behaviours/instances/conservation_exploration/RecordCreaturesBehavior.java
@@ -1,4 +1,4 @@
-package com.lovetropics.minigames.common.minigames.behaviours.instances.collaborative_conservation;
+package com.lovetropics.minigames.common.minigames.behaviours.instances.conservation_exploration;
 
 import com.lovetropics.lib.entity.FireworkUtil;
 import com.lovetropics.minigames.common.item.MinigameItems;
